+++ sdk/attestation/azure-security-attestation/inc/azure/attestation/attestation_client.hpp
@@ -248,7 +248,6 @@ namespace Azure { namespace Security { namespace Attestation {
   private:
     Azure::Core::Url m_endpoint;
     std::string m_apiVersion;
-    std::shared_ptr<Azure::Core::Credentials::TokenCredential const> m_credentials;
     std::shared_ptr<Azure::Core::Http::_internal::HttpPipeline> m_pipeline;
     AttestationTokenValidationOptions m_tokenValidationOptions;
     std::vector<Models::AttestationSigner> m_attestationSigners;