+++ sdk/attestation/azure-security-attestation/src/attestation_client_options.cpp
@@ -4,9 +4,8 @@
 #include "azure/attestation/attestation_client_options.hpp"
 
 namespace Azure { namespace Security { namespace Attestation {
-  const ServiceVersion ServiceVersion::V2020_10_01("2020-10-01");
 
-  const AttestationDataType AttestationDataType ::Binary("Binary");
+  const AttestationDataType AttestationDataType::Binary("Binary");
   const AttestationDataType AttestationDataType::Json("JSON");
 
 }}} // namespace Azure::Security::Attestation