+++ sdk/core/azure-core-tracing-opentelemetry/test/ut/azure_core_otel_test.cpp
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #define USE_MEMORY_EXPORTER 1
+#include "../src/opentelemetry_private.hpp"
 #include "azure/core/tracing/opentelemetry/opentelemetry.hpp"
 #include <azure/core/test/test_base.hpp>
 
@@ -137,18 +138,24 @@ protected:
 
 TEST_F(OpenTelemetryTests, Basic)
 {
-  // Simple create an OTel telemetry provider as a static member variable.
+  // Simple create an OTel telemetry provider and call a method on the concrete implementation.
   {
-    Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider provider;
-    auto tracer = provider.CreateTracer("TracerName", "1.0");
+    auto provider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create();
+
+    auto tracer
+        = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(provider)
+              ->CreateTracer("TracerName", "1.0");
     EXPECT_TRUE(tracer);
   }
 
-  // Create a shared provider using the tracing abstract classes.
+  // Create a provider using the tracing abstract classes.
   {
-    std::shared_ptr<Azure::Core::Tracing::TracerProvider> provider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>();
-    auto tracer = provider->CreateTracer("TracerName", "1.0");
+    auto otProvider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create();
+    auto provider = otProvider;
+
+    std::shared_ptr<Azure::Core::Tracing::_internal::TracerProviderImpl> providerImpl
+        = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(provider);
+    auto tracer = providerImpl->CreateTracer("TracerName", "1.0");
     EXPECT_TRUE(tracer);
   }
 
@@ -157,9 +164,11 @@ TEST_F(OpenTelemetryTests, Basic)
     auto rawTracer(opentelemetry::trace::Provider::GetTracerProvider());
 
     auto traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(rawTracer);
+        = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(rawTracer);
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", "1.0");
     EXPECT_TRUE(tracer);
   }
 
@@ -168,10 +177,11 @@ TEST_F(OpenTelemetryTests, Basic)
   {
     auto otelProvider(CreateOpenTelemetryProvider());
     auto traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-            otelProvider);
+        = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(otelProvider);
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
   }
 }
 
@@ -179,28 +189,32 @@ TEST_F(OpenTelemetryTests, CreateSpanSimple)
 {
   // Simple create an OTel telemetry provider as a static member variable.
   {
-    Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider provider;
-    auto tracer = provider.CreateTracer("TracerName", "1.0");
+    auto provider(Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create());
+    auto tracer
+        = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(provider)
+              ->CreateTracer("TracerName", "1.0");
     EXPECT_TRUE(tracer);
-    auto span = tracer->CreateSpan("My Span");
+    auto span = tracer->CreateSpan("My Span", {});
     EXPECT_TRUE(span);
 
-    span->End();
+    span->End({});
   }
 
   // Create a provider using the OpenTelemetry reference provider (this will be a working provider
   // using the ostream logger).
   {
-    std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-            CreateOpenTelemetryProvider());
+    auto otProvider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
+        CreateOpenTelemetryProvider());
+    std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider(otProvider);
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
     {
-      auto span = tracer->CreateSpan("My Span2");
+      auto span = tracer->CreateSpan("My Span2", {});
       EXPECT_TRUE(span);
 
-      span->End();
+      span->End({});
     }
     // Return the collected spans.
     auto spans = m_spanData->GetSpans();
@@ -307,31 +321,38 @@ TEST_F(OpenTelemetryTests, CreateSpanWithOptions)
 {
   // Simple create an OTel telemetry provider as a static member variable.
   {
-    Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider provider;
-    auto tracer = provider.CreateTracer("TracerName", "1.0");
+    auto provider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
+        opentelemetry::trace::Provider::GetTracerProvider());
+    auto tracer
+        = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(provider)
+              ->CreateTracer("TracerName", "1.0");
     EXPECT_TRUE(tracer);
     Azure::Core::Tracing::_internal::CreateSpanOptions options;
     auto span = tracer->CreateSpan("My Span", options);
     EXPECT_TRUE(span);
 
-    span->End();
+    span->End({});
   }
 
   // Create a provider using the OpenTelemetry reference provider (this will be a working provider
   // using the ostream logger).
   {
-    std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
+    std::shared_ptr<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider> traceProvider
+        = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
             CreateOpenTelemetryProvider());
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
+    auto tracerImpl
+        = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+            traceProvider);
+
+    auto tracer = tracerImpl->CreateTracer("TracerName", {});
     {
       Azure::Core::Tracing::_internal::CreateSpanOptions options;
       options.Kind = Azure::Core::Tracing::_internal::SpanKind::Client;
       auto span = tracer->CreateSpan("Client Span", options);
       EXPECT_TRUE(span);
 
-      span->End();
+      span->End({});
     }
     {
       Azure::Core::Tracing::_internal::CreateSpanOptions options;
@@ -339,7 +360,7 @@ TEST_F(OpenTelemetryTests, CreateSpanWithOptions)
       auto span = tracer->CreateSpan("Consumer Span", options);
       EXPECT_TRUE(span);
 
-      span->End();
+      span->End({});
     }
     {
       Azure::Core::Tracing::_internal::CreateSpanOptions options;
@@ -347,7 +368,7 @@ TEST_F(OpenTelemetryTests, CreateSpanWithOptions)
       auto span = tracer->CreateSpan("Internal Span", options);
       EXPECT_TRUE(span);
 
-      span->End();
+      span->End({});
     }
     {
       Azure::Core::Tracing::_internal::CreateSpanOptions options;
@@ -355,7 +376,7 @@ TEST_F(OpenTelemetryTests, CreateSpanWithOptions)
       auto span = tracer->CreateSpan("Producer Span", options);
       EXPECT_TRUE(span);
 
-      span->End();
+      span->End({});
     }
     {
       Azure::Core::Tracing::_internal::CreateSpanOptions options;
@@ -363,12 +384,7 @@ TEST_F(OpenTelemetryTests, CreateSpanWithOptions)
       auto span = tracer->CreateSpan("Server Span", options);
       EXPECT_TRUE(span);
 
-      span->End();
-    }
-    {
-      Azure::Core::Tracing::_internal::CreateSpanOptions options;
-      options.Kind = Azure::Core::Tracing::_internal::SpanKind("Bogus");
-      EXPECT_THROW(tracer->CreateSpan("Bogus Span", options), std::runtime_error);
+      span->End({});
     }
     // Return the collected spans.
     auto spans = m_spanData->GetSpans();
@@ -390,11 +406,13 @@ TEST_F(OpenTelemetryTests, CreateSpanWithOptions)
     // Create a provider using the OpenTelemetry reference provider (this will be a working provider
     // using the ostream logger).
     {
-      std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-          = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-              CreateOpenTelemetryProvider());
+      auto otProvider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
+          CreateOpenTelemetryProvider());
+      std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider(otProvider);
 
-      auto tracer = traceProvider->CreateTracer("TracerName");
+      auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                        traceProvider)
+                        ->CreateTracer("TracerName", {});
       {
         Azure::Core::Tracing::_internal::CreateSpanOptions options;
         options.Attributes = std::make_unique<
@@ -404,7 +422,7 @@ TEST_F(OpenTelemetryTests, CreateSpanWithOptions)
         auto span = tracer->CreateSpan("Client Span", options);
         EXPECT_TRUE(span);
 
-        span->End();
+        span->End({});
 
         // Return the collected spans.
         auto spans = m_spanData->GetSpans();
@@ -429,11 +447,13 @@ TEST_F(OpenTelemetryTests, NestSpans)
 
   {
     std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
+        = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
             CreateOpenTelemetryProvider());
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
-    auto span = tracer->CreateSpan("SpanOuter");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
+    auto span = tracer->CreateSpan("SpanOuter", {});
     EXPECT_TRUE(span);
     {
       Azure::Core::Tracing::_internal::CreateSpanOptions so;
@@ -444,19 +464,19 @@ TEST_F(OpenTelemetryTests, NestSpans)
       // Span 3's parent is SpanOuter.
       so.ParentSpan = span;
       auto span4 = tracer->CreateSpan("SpanInner4", so);
-      span2->End();
+      span2->End({});
 
-      span->End();
-      span4->End();
-      span3->End();
+      span->End({});
+      span4->End({});
+      span3->End({});
     }
     {
       Azure::Core::Tracing::_internal::CreateSpanOptions so;
       so.ParentSpan = span;
       auto span5 = tracer->CreateSpan("SequentialInner", so);
       auto span6 = tracer->CreateSpan("SequentialInner2", so);
-      span5->End();
-      span6->End();
+      span5->End({});
+      span6->End({});
     }
 
     // Return the collected spans.
@@ -500,17 +520,19 @@ TEST_F(OpenTelemetryTests, SetStatus)
 
   {
     std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
+        = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
             CreateOpenTelemetryProvider());
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
-    auto span = tracer->CreateSpan("StatusSpan");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
+    auto span = tracer->CreateSpan("StatusSpan", {});
     EXPECT_TRUE(span);
 
-    span->SetStatus(Azure::Core::Tracing::_internal::SpanStatus::Error);
-    span->SetStatus(Azure::Core::Tracing::_internal::SpanStatus::Ok);
+    span->SetStatus(Azure::Core::Tracing::_internal::SpanStatus::Error, {});
+    span->SetStatus(Azure::Core::Tracing::_internal::SpanStatus::Ok, {});
 
-    span->End();
+    span->End({});
 
     // Return the collected spans.
     auto spans = m_spanData->GetSpans();
@@ -520,17 +542,18 @@ TEST_F(OpenTelemetryTests, SetStatus)
   }
 
   {
-    std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-            CreateOpenTelemetryProvider());
+    auto traceProvider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
+        CreateOpenTelemetryProvider());
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
-    auto span = tracer->CreateSpan("StatusSpan");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
+    auto span = tracer->CreateSpan("StatusSpan", {});
     EXPECT_TRUE(span);
 
     span->SetStatus(Azure::Core::Tracing::_internal::SpanStatus::Error, "Something went wrong.");
 
-    span->End();
+    span->End({});
 
     // Return the collected spans.
     auto spans = m_spanData->GetSpans();
@@ -542,17 +565,18 @@ TEST_F(OpenTelemetryTests, SetStatus)
 
   // Set to Unset.
   {
-    std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-            CreateOpenTelemetryProvider());
+    auto traceProvider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
+        CreateOpenTelemetryProvider());
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
-    auto span = tracer->CreateSpan("StatusSpan");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
+    auto span = tracer->CreateSpan("StatusSpan", {});
     EXPECT_TRUE(span);
 
-    span->SetStatus(Azure::Core::Tracing::_internal::SpanStatus::Unset);
+    span->SetStatus(Azure::Core::Tracing::_internal::SpanStatus::Unset, {});
 
-    span->End();
+    span->End({});
 
     // Return the collected spans.
     auto spans = m_spanData->GetSpans();
@@ -563,15 +587,16 @@ TEST_F(OpenTelemetryTests, SetStatus)
 
   // Not set.
   {
-    std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-            CreateOpenTelemetryProvider());
+    auto traceProvider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
+        CreateOpenTelemetryProvider());
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
-    auto span = tracer->CreateSpan("StatusSpan");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
+    auto span = tracer->CreateSpan("StatusSpan", {});
     EXPECT_TRUE(span);
 
-    span->End();
+    span->End({});
 
     // Return the collected spans.
     auto spans = m_spanData->GetSpans();
@@ -582,16 +607,14 @@ TEST_F(OpenTelemetryTests, SetStatus)
 
   // Invalid status.
   {
-    std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-            CreateOpenTelemetryProvider());
+    auto traceProvider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
+        CreateOpenTelemetryProvider());
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
-    auto span = tracer->CreateSpan("StatusSpan");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
+    auto span = tracer->CreateSpan("StatusSpan", {});
     EXPECT_TRUE(span);
-
-    EXPECT_THROW(
-        span->SetStatus(Azure::Core::Tracing::_internal::SpanStatus("Bogus")), std::runtime_error);
 
     // Return the collected spans.
     auto spans = m_spanData->GetSpans();
@@ -603,12 +626,13 @@ TEST_F(OpenTelemetryTests, AddSpanAttributes)
 {
 
   {
-    std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-            CreateOpenTelemetryProvider());
+    auto traceProvider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
+        CreateOpenTelemetryProvider());
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
-    auto span = tracer->CreateSpan("AttributeSpan");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
+    auto span = tracer->CreateSpan("AttributeSpan", {});
     EXPECT_TRUE(span);
 
     Azure::Core::Tracing::OpenTelemetry::_detail::OpenTelemetryAttributeSet attributeSet;
@@ -622,7 +646,7 @@ TEST_F(OpenTelemetryTests, AddSpanAttributes)
     std::string stringValue("std::string.");
     attributeSet.AddAttribute("stdstring", stringValue);
     span->AddAttributes(attributeSet);
-    span->End();
+    span->End({});
 
     // Return the collected spans.
     auto spans = m_spanData->GetSpans();
@@ -643,12 +667,13 @@ TEST_F(OpenTelemetryTests, AddSpanAttributes)
 TEST_F(OpenTelemetryTests, AddSpanEvents)
 {
   {
-    std::shared_ptr<Azure::Core::Tracing::TracerProvider> traceProvider
-        = std::make_shared<Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider>(
-            CreateOpenTelemetryProvider());
+    auto traceProvider = Azure::Core::Tracing::OpenTelemetry::OpenTelemetryProvider::Create(
+        CreateOpenTelemetryProvider());
 
-    auto tracer = traceProvider->CreateTracer("TracerName");
-    auto span = tracer->CreateSpan("SpanWithEvents");
+    auto tracer = Azure::Core::Tracing::_internal::TracerProviderImplGetter::TracerImplFromTracer(
+                      traceProvider)
+                      ->CreateTracer("TracerName", {});
+    auto span = tracer->CreateSpan("SpanWithEvents", {});
     EXPECT_TRUE(span);
 
     span->AddEvent("String Event");
@@ -667,7 +692,7 @@ TEST_F(OpenTelemetryTests, AddSpanEvents)
       attributeSet.AddAttribute("stdstring", stringValue);
       span->AddEvent("Event With Attributes", attributeSet);
 
-      span->End();
+      span->End({});
 
       // Return the collected spans.
       auto spans = m_spanData->GetSpans();