+++ sdk/keyvault/azure-security-keyvault-keys/test/ut/key_client_base_test.hpp
@@ -9,12 +9,11 @@
 #include <gtest/gtest.h>
 
 #include <azure/core/context.hpp>
+#include <azure/core/test/test_base.hpp>
 #include <azure/core/uuid.hpp>
 #include <azure/identity/client_secret_credential.hpp>
 #include <azure/keyvault/keys.hpp>
 #include <azure/keyvault/keys/cryptography/cryptography_client.hpp>
-
-#include <azure/core/test/test_base.hpp>
 
 #include <chrono>
 #include <cstdio>
@@ -33,6 +32,16 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys { nam
     std::string m_keyVaultHsmUrl;
     int m_testPollingTimeOutMinutes = 20;
     std::chrono::milliseconds m_testPollingIntervalMs = std::chrono::minutes(1);
+
+    // Reads the current test instance name.
+    // Name gets also sanitized (special chars are removed) to avoid issues when recording or
+    // creating. This also return the name with suffix if the "AZURE_LIVE_TEST_SUFFIX" exists.
+    std::string GetTestName(bool sanitize = true)
+    {
+      auto output = m_keyVaultUrl.compare(m_keyVaultHsmUrl) == 0 ? "Same" : "NotSame";
+      std::cout << "\n Keyvault and HSM are" << output;
+      return Azure::Core::Test::TestBase::GetTestNameSuffix(sanitize);
+    }
 
     Azure::Security::KeyVault::Keys::KeyClient const& GetClientForTest(std::string const& testName)
     {
@@ -71,13 +80,13 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys { nam
       UpdateWaitingTime(m_testPollingIntervalMs);
     }
 
-    void CreateHsmClient()
+    void CreateHsmClient(std::string hsmUrl = "")
     {
       KeyClientOptions options;
       m_client = InitTestClient<
           Azure::Security::KeyVault::Keys::KeyClient,
           Azure::Security::KeyVault::Keys::KeyClientOptions>(
-          m_keyVaultHsmUrl, m_credential, options);
+          hsmUrl.length() == 0 ? m_keyVaultHsmUrl : hsmUrl, m_credential, options);
     }
 
   public: