+++ sdk/keyvault/azure-security-keyvault-keys/test/ut/key_client_test.cpp
@@ -3,6 +3,8 @@
 
 #include "gtest/gtest.h"
 
+#include "private/key_constants.hpp"
+#include "private/key_serializers.hpp"
 #include <azure/core/context.hpp>
 #include <azure/identity/client_secret_credential.hpp>
 #include <azure/keyvault/keys.hpp>
@@ -30,18 +32,10 @@ TEST(KeyVaultKeyClientUnitTest, ServiceVersion)
 {
   auto credential
       = std::make_shared<Azure::Identity::ClientSecretCredential>("tenantID", "AppId", "SecretId");
-  {
-    // 7.2
-    EXPECT_NO_THROW(auto options = KeyClientOptions(ServiceVersion::V7_2);
-                    KeyClient keyClient("http://account.vault.azure.net", credential, options);
-                    EXPECT_EQ(options.Version.ToString(), "7.2"););
-  }
-  {
-    // arbitrary version
-    EXPECT_NO_THROW(auto options = KeyClientOptions(ServiceVersion("1.0"));
-                    KeyClient keyClient("http://account.vault.azure.net", credential, options);
-                    EXPECT_EQ(options.Version.ToString(), "1.0"););
-  }
+  // 7.3
+  EXPECT_NO_THROW(auto options = KeyClientOptions();
+                  KeyClient keyClient("http://account.vault.azure.net", credential, options);
+                  EXPECT_EQ(options.ApiVersion, "7.3"););
 }
 
 TEST(KeyVaultKeyClientUnitTest, GetUrl)
@@ -52,4 +46,64 @@ TEST(KeyVaultKeyClientUnitTest, GetUrl)
   auto url = "vaultUrl";
   KeyClient keyClient(url, credential);
   EXPECT_EQ(url, keyClient.GetUrl());
+}
+
+TEST(KeyReleaseOptionsUnitTest, None)
+{
+  KeyReleaseOptions options;
+  auto serialized = _detail::KeyReleaseOptionsSerializer::KeyReleaseOptionsSerialize(options);
+  auto deserialized = Azure::Core::Json::_internal::json::parse(serialized);
+
+  EXPECT_EQ(options.Target, deserialized[_detail::TargetValue]);
+  EXPECT_EQ(nullptr, deserialized[_detail::NonceValue]);
+  EXPECT_EQ(nullptr, deserialized[_detail::EncryptionValue]);
+}
+
+TEST(KeyReleaseOptionsUnitTest, One)
+{
+  KeyReleaseOptions options;
+  options.Target = "xyz";
+  auto serialized = _detail::KeyReleaseOptionsSerializer::KeyReleaseOptionsSerialize(options);
+  auto deserialized = Azure::Core::Json::_internal::json::parse(serialized);
+
+  EXPECT_EQ(options.Target, deserialized[_detail::TargetValue]);
+  EXPECT_EQ(nullptr, deserialized[_detail::NonceValue]);
+  EXPECT_EQ(nullptr, deserialized[_detail::EncryptionValue]);
+}
+
+TEST(KeyReleaseOptionsUnitTest, Most)
+{
+  KeyReleaseOptions options;
+  options.Target = "xyz";
+  options.Nonce = "abc";
+  auto serialized = _detail::KeyReleaseOptionsSerializer::KeyReleaseOptionsSerialize(options);
+  auto deserialized = Azure::Core::Json::_internal::json::parse(serialized);
+
+  EXPECT_EQ(options.Target, deserialized[_detail::TargetValue]);
+  EXPECT_EQ(options.Nonce.Value(), deserialized[_detail::NonceValue]);
+  EXPECT_EQ(nullptr, deserialized[_detail::EncryptionValue]);
+}
+
+TEST(KeyReleaseOptionsUnitTest, All)
+{
+  KeyReleaseOptions options;
+  options.Target = "xyz";
+  options.Nonce = "abc";
+  options.Encryption = KeyEncryptionAlgorithm::CKM_RSA_AES_KEY_WRAP;
+  auto serialized = _detail::KeyReleaseOptionsSerializer::KeyReleaseOptionsSerialize(options);
+  auto deserialized = Azure::Core::Json::_internal::json::parse(serialized);
+
+  EXPECT_EQ(options.Target, deserialized[_detail::TargetValue]);
+  EXPECT_EQ(options.Nonce.Value(), deserialized[_detail::NonceValue]);
+  EXPECT_EQ(options.Encryption.Value().ToString(), deserialized[_detail::EncryptionValue]);
+}
+
+TEST(KeyEncryptionAlgorithmUnitTest, CheckValues)
+{
+  EXPECT_EQ(
+      KeyEncryptionAlgorithm::CKM_RSA_AES_KEY_WRAP.ToString(), _detail::CKM_RSA_AES_KEY_WRAP_Value);
+  EXPECT_EQ(
+      KeyEncryptionAlgorithm::RSA_AES_KEY_WRAP_256.ToString(), _detail::RSA_AES_KEY_WRAP_256_Value);
+  EXPECT_EQ(
+      KeyEncryptionAlgorithm::RSA_AES_KEY_WRAP_384.ToString(), _detail::RSA_AES_KEY_WRAP_384_Value);
 }