+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_container_client.hpp
@@ -12,6 +12,7 @@
 namespace Azure { namespace Storage { namespace Blobs {
 
   class BlobLeaseClient;
+  class BlobContainerBatch;
 
   /**
    * The BlobContainerClient allows you to manipulate Azure Storage containers and their
@@ -284,25 +285,41 @@ namespace Azure { namespace Storage { namespace Blobs {
         const UploadBlockBlobOptions& options = UploadBlockBlobOptions(),
         const Azure::Core::Context& context = Azure::Core::Context()) const;
 
+    /**
+     * @brief Creates a new batch object to collect subrequests that can be submitted together via
+     * SubmitBatch.
+     *
+     * @return A new batch object.
+     */
+    BlobContainerBatch CreateBatch() const;
+
+    /**
+     * @brief Submits a batch of subrequests.
+     *
+     * @param batch The batch object containing subrequests.
+     * @param options Optional parameters to execute this function.
+     * @param context Context for cancelling long running operations.
+     * @return A SubmitBlobBatchResult.
+     * @remark This function will throw only if there's something wrong with the batch request
+     * (parent request).
+     */
+    Response<Models::SubmitBlobBatchResult> SubmitBatch(
+        const BlobContainerBatch& batch,
+        const SubmitBlobBatchOptions& options = SubmitBlobBatchOptions(),
+        const Core::Context& context = Core::Context()) const;
+
   private:
     Azure::Core::Url m_blobContainerUrl;
     std::shared_ptr<Azure::Core::Http::_internal::HttpPipeline> m_pipeline;
     Azure::Nullable<EncryptionKey> m_customerProvidedKey;
     Azure::Nullable<std::string> m_encryptionScope;
 
-    explicit BlobContainerClient(
-        Azure::Core::Url blobContainerUrl,
-        std::shared_ptr<Azure::Core::Http::_internal::HttpPipeline> pipeline,
-        Azure::Nullable<EncryptionKey> customerProvidedKey,
-        Azure::Nullable<std::string> encryptionScope)
-        : m_blobContainerUrl(std::move(blobContainerUrl)), m_pipeline(std::move(pipeline)),
-          m_customerProvidedKey(std::move(customerProvidedKey)),
-          m_encryptionScope(std::move(encryptionScope))
-    {
-    }
+    std::shared_ptr<Azure::Core::Http::_internal::HttpPipeline> m_batchRequestPipeline;
+    std::shared_ptr<Azure::Core::Http::_internal::HttpPipeline> m_batchSubrequestPipeline;
 
     friend class BlobServiceClient;
     friend class BlobLeaseClient;
+    friend class BlobContainerBatch;
   };
 
 }}} // namespace Azure::Storage::Blobs